% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flake_template_files.R
\name{flake_template_files}
\alias{flake_template_files}
\title{Return the path to a template of current NML files}
\usage{
flake_template_files()
}
\description{
This returns a path to a directory with example NML files for running FLake and DAT files which contain the model forcing data.
}
\examples{
\dontrun{
template_files()
}

}
\author{
Tadhg Moore
}
\keyword{methods}
