% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_flake.R
\name{run_flake}
\alias{run_flake}
\title{Run the FLake model}
\usage{
run_flake(sim_folder = ".", nml_file = NULL, verbose = FALSE)
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{nml_file}{filepath; to file with FLake setup. Defaults to 'flake.nml'}

\item{verbose}{Save output as character vector. Defaults to FALSE}
}
\description{
This runs the FLake model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain valid NML files.
}
\examples{
sim_folder <- system.file('extdata', package = 'FLakeR')
run_flake(sim_folder, nml_file = 'Heiligensee80-96.nml')
}
\author{
Tadhg Moore
}
\keyword{methods}
